requirements.txt
asgiref==3.6.0
astroid==2.15.0
autopep8==2.0.2
boto3==1.26.93
botocore==1.29.93
certifi==2022.6.15
dill==0.3.6
distlib==0.3.6
dj-database-url==1.2.0
Django==4.1.7
django-environ==0.10.0
django-on-heroku==1.1.2
djangorestframework==3.14.0
filelock==3.8.0
gunicorn==20.1.0
isort==5.12.0
jmespath==1.0.1
lazy-object-proxy==1.9.0
mccabe==0.7.0
pipenv==2022.8.24
platformdirs==2.5.2
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pylint==2.16.4
python-dateutil==2.8.2
pytz==2022.7.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tomlkit==0.11.6
typing_extensions==4.5.0
urllib3==1.26.15
virtualenv==20.16.3
virtualenv-clone==0.5.7
whitenoise==6.4.0
wrapt==1.15.0
